Refatora Selected Work para otimizar acessibilidade e performance
